Fix JsonAddressBookStorageTest
