+++ src/test/java/seedu/address/storage/JsonAddressBookStorageTest.java
@@ -3,10 +3,13 @@ package seedu.address.storage;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static seedu.address.testutil.Assert.assertThrows;
-import static seedu.address.testutil.TypicalPersons.ALICE;
-import static seedu.address.testutil.TypicalPersons.HOON;
-import static seedu.address.testutil.TypicalPersons.IDA;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalAddressBook.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalBuyers.ALICE;
+import static seedu.address.testutil.TypicalBuyers.HOON;
+import static seedu.address.testutil.TypicalBuyers.IDA;
+import static seedu.address.testutil.TypicalSellers.SALICE;
+import static seedu.address.testutil.TypicalSellers.SHOON;
+import static seedu.address.testutil.TypicalSellers.SIDA;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -72,14 +75,17 @@ public class JsonAddressBookStorageTest {
         assertEquals(original, new AddressBook(readBack));
 
         // Modify data, overwrite exiting file, and read back
-        original.addPerson(HOON);
-        original.removePerson(ALICE);
+        original.addBuyer(HOON);
+        original.addSeller(SHOON);
+        original.removeBuyer(ALICE);
+        original.removeSeller(SALICE);
         jsonAddressBookStorage.saveAddressBook(original, filePath);
         readBack = jsonAddressBookStorage.readAddressBook(filePath).get();
         assertEquals(original, new AddressBook(readBack));
 
         // Save and read without specifying file path
-        original.addPerson(IDA);
+        original.addBuyer(IDA);
+        original.addSeller(SIDA);
         jsonAddressBookStorage.saveAddressBook(original); // file path not specified
         readBack = jsonAddressBookStorage.readAddressBook().get(); // file path not specified
         assertEquals(original, new AddressBook(readBack));
