+++ src/test/java/seedu/address/storage/JsonSerializableAddressBookTest.java
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.JsonUtil;
 import seedu.address.model.AddressBook;
-import seedu.address.testutil.TypicalPersons;
 
 public class JsonSerializableAddressBookTest {
